# server/swagger.yaml
...
openapi: 3.0.0
info:
  title: MyUber
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /booking/check:
    post:
      summary: Check if user can book a car and a driver for his movement
      security: []   # No security
      tags:
        - Booking
      parameters:
        - name: endPlaceLat
          in: path
          description: 
          required: true
          schema:
            type: string
        - name: endPlaceLng
          in: path
          description: 
          required: true
          schema:
            type: number
        - name: endTime
          in: path
          description: 
          required: true
          schema:
            type: integer
        - name: startPlaceLat
          in: path
          description: 
          required: true
          schema:
            type: number
        - name: startPlaceLng
          in: path
          description: 
          required: true
          schema:
            type: number
        - name: startTime
          in: path
          description: 
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
...
  /booking/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Booking
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Booking
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    post:
      summary: CRUD ACTION update
      tags:
        - Booking
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Booking"
    
  #-----  Car  -----
...
  /car/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Car
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Car
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
    post:
      summary: CRUD ACTION update
      tags:
        - Car
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Car"
  /car/findByCapacity/{key}:
    get:
      summary: CRUD ACTION findByCapacity
      tags:
        - Car
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
    get:
      summary: CRUD ACTION findByCapacity
      tags:
        - Car
      parameters:
        - name: Status
          in: path
          description: Status of the Car
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
...
  /user/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
    
  #-----  User  -----
  /Users/{id}/changePassword:
    post:
      summary: Change password by admin API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordAdmin:
                  type: string
      responses:
        '200':
          description: OK
  /Users/{id}:
    post:
      summary: Update User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    delete:
      summary: Delete User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Booking:
      properties:
        _id:
          type: string
        CarID:
          type: string
        DriverID:
          type: string
        EndPlaceLat:
          type: number
        EndPlaceLng:
          type: number
        EndTime:
          type: integer
        StartPlaceLat:
          type: number
        StartPlaceLng:
          type: number
        StartTime:
          type: integer
        Status:
          type: integer
      required:
        -  _id
        - Status
    Car:
      properties:
        _id:
          type: string
        BrandID:
          type: string
        Capacity:
          type: integer
        ModelID:
          type: string
        Number:
          type: string
        Status:
          type: integer
      required:
        -  _id
        - BrandID
        - Capacity
        - ModelID
        - Number
        - Status
...
    User:
      properties:
        _id:
          type: string
        mail:
          type: string
        name:
          type: string
        password:
          type: string
        roles:
          type: string
        surname:
          type: string
        username:
          type: string
      required:
        -  _id
        - password
        - username